requirements.txt
joblib==1.0.1
matplotlib==3.3.4
mpose
numpy==1.19.5
optuna==2.8.0
pandas==1.1.5
PyYAML==6.0
scikit_learn==1.0.2
scipy==1.5.4
tensorflow==2.6.0
tensorflow_addons==0.14.0
tensorflow_datasets==4.4.0
keras ==2.6.0  # We get some error about dtensor https://stackoverflow.com/questions/72255562/cannot-import-name-dtensor-from-tensorflow-compat-v2-experimental
# Install pytorch with cuda11.8. Need to find a version that works for both. Atm cuda should be V11.2.152 according to
# docker for tf2.6. Also this current version of pt req typing-extensions~=4.9.0 but tf2.6 requires typing-extensions~=3.7.4
pytorch-model-summary==0.1.2
pytorch-ignite==0.4.13
--extra-index-url https://download.pytorch.org/whl/cu113
pytorch-ignite==0.4.13
torch==1.12.1+cu113 # Was 2.1.1+cu118
torchvision==0.13.1+cu113
torchaudio==0.12.1